+++ text2bin.cpp
@@ -11,7 +11,7 @@ int main(int, char **argv) {
     ifstream fin(argv[1], ios_base::in);
     ofstream fout(argv[2], ios_base::out | ios_base::binary);
 
-    int value = 0;
+    uint64_t value = 0;
     while (fin) {
         fin >> value;
         if (!fin) {
